requirements.txt
nats-py>=2.6.0
protobuf==5.29.3
python-dotenv>=1.0.0
influxdb-client>=1.40.0
fastapi>=0.110.0
uvicorn>=0.27.0
strawberry-graphql>=0.220.0
discord.py>=2.3.0